+++ src/Utility/functions.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 namespace Miklcct\ThinPhpApp\Utility;
 
 /**
- * Run a callback on an object if it is not NULL
+ * Run a callback on a value if it is not NULL
  *
- * @param $object
- * @param callable $callback
+ * @param mixed|null $value A value which may possibly be NULL
+ * @param callable $callback Called when $object is not NULL
+ * @param mixed|null $default Value returned when $object is NULL
  * @return mixed|null
  */
-function nullable($object, callable $callback) {
-    return $object !== NULL ? $callback($object) : NULL;
+function nullable($value, callable $callback, $default = NULL) {
+    return $value !== NULL ? $callback($value) : $default;
 }